requirements.txt
matplotlib==3.5.1
numpy==1.21.5
Pillow==9.0.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
SimpleITK==2.1.1
sklearn==0.0
tensorboard==2.2.0
tensorboardX==2.5
timm==0.9.12
torch==1.10.1
torchvision==0.11.2+cu113
tqdm==4.62.3

